[aerial_robot_model] fix CmakeList: depend on casadi_eigen and separate pinocchio model library and execution node
